move the visualizer to mic and use chunks from mic
